+++ bw.workflow.citizenenrollment/src/main/java/sa/gov/nic/bio/bw/workflow/citizenenrollment/controllers/IrisCapturingFxController.java
@@ -40,18 +40,13 @@ public class IrisCapturingFxController extends WizardStepFxControllerBase {
     private static final int RIGHT_IRIS_ONLY = 1;
     private static final int LEFT_AND_RIGHT_IRIS = 2;
 
-    @Input
-    private Boolean hidePreviousButton;
-    @Input
-    private Boolean hideStartOverButton;
-    @Input
-    private Boolean hideSkipButton;
-    @Output
-    private String capturedRightIrisBase64;
-    @Output
-    private String capturedLeftIrisBase64;
-    @Output
-    private Boolean Skip;
+    @Input private Boolean hidePreviousButton;
+    @Input private Boolean hideStartOverButton;
+    @Input private Boolean hideSkipButton;
+    @Output private String capturedRightIrisBase64;
+    @Output private String capturedLeftIrisBase64;
+    @Output private String capturedRightIrisCompressedBase64;
+    @Output private String capturedLeftIrisCompressedBase64;
 
     @FXML
     private VBox paneControlsInnerContainer;
@@ -317,8 +312,6 @@ public class IrisCapturingFxController extends WizardStepFxControllerBase {
     @FXML
     private void onCaptureIrisButtonClicked(ActionEvent event) {
 
-        Skip = false;
-
         capturedRightIrisBase64 = null;
         capturedLeftIrisBase64 = null;
         ivCapturedFirstIris.setImage(null);
@@ -360,8 +353,14 @@ public class IrisCapturingFxController extends WizardStepFxControllerBase {
                 CaptureIrisResponse result = taskResponse.getResult();
 
                 if (result.getReturnCode() == CaptureIrisResponse.SuccessCodes.SUCCESS) {
-                    if (cbSkippedRightIris.isSelected()) { capturedRightIrisBase64 = result.getRightIrisImageBase64(); }
-                    if (cbSkippedLeftIris.isSelected()) { capturedLeftIrisBase64 = result.getLeftIrisImageBase64(); }
+                    if(cbSkippedRightIris.isSelected()) {
+                        capturedRightIrisBase64 = result.getRightIrisImageBase64();
+                        capturedRightIrisCompressedBase64 = result.getRightIrisCompressedImageBase64();
+                    }
+                    if(cbSkippedLeftIris.isSelected()) {
+                        capturedLeftIrisBase64 = result.getLeftIrisImageBase64();
+                        capturedLeftIrisCompressedBase64 = result.getLeftIrisCompressedImageBase64();
+                    }
                     showIris();
                 }
                 else {
@@ -464,7 +463,10 @@ public class IrisCapturingFxController extends WizardStepFxControllerBase {
 
     @FXML
     private void onSkipButtonClicked() {
-        Skip = true;
+        capturedRightIrisBase64 = null;
+        capturedLeftIrisBase64 = null;
+        capturedRightIrisCompressedBase64 = null;
+        capturedLeftIrisCompressedBase64 = null;
         goNext();
 
     }