+++ bw.workflow.citizenenrollment/src/main/java/sa/gov/nic/bio/bw/workflow/citizenenrollment/controllers/RegisteringCitizenPaneFxController.java
@@ -10,31 +10,21 @@ import javafx.scene.image.ImageView;
 import sa.gov.nic.bio.bw.core.Context;
 import sa.gov.nic.bio.bw.core.controllers.WizardStepFxControllerBase;
 import sa.gov.nic.bio.bw.core.utils.FxmlFile;
+import sa.gov.nic.bio.bw.core.utils.GuiUtils;
 import sa.gov.nic.bio.bw.core.workflow.Input;
 import sa.gov.nic.bio.bw.core.workflow.Output;
-import sa.gov.nic.bio.bw.workflow.citizenenrollment.tasks.CheckCitizenRegistrationWorkflowTask;
-import sa.gov.nic.bio.bw.workflow.citizenenrollment.tasks.CheckIrisRegistrationWorkflowTask;
+import sa.gov.nic.bio.bw.workflow.citizenenrollment.tasks.CheckCitizenRegistrationWorkflowTask.Status;
 
 @FxmlFile("registeringCitizen.fxml")
 public class RegisteringCitizenPaneFxController extends WizardStepFxControllerBase {
     public enum Request {
         SUBMIT_CITIZEN_REGISTRATION,
-        CHECK_CITIZEN_REGISTRATION,
-        SUBMIT_IRIS_REGISTRATION,
-        CHECK_IRIS_REGISTRATION
+        CHECK_CITIZEN_REGISTRATION
     }
 
-    @Input
-    private CheckCitizenRegistrationWorkflowTask.Status citizenRegistrationStatus;
-    @Input
-    private CheckIrisRegistrationWorkflowTask.Status irisRegistrationStatus;
-
-    @Input(alwaysRequired = true)
-    private Boolean skipIris;
-    @Input
-    private Boolean isEnrollmentProcessStart;
-    @Output
-    private Request request;
+    @Input private Status citizenRegistrationStatus;
+    @Input private Boolean isEnrollmentProcessStart;
+    @Output private Request request;
 
     @FXML
     private ProgressIndicator CPiProgress;
@@ -44,15 +34,6 @@ public class RegisteringCitizenPaneFxController extends WizardStepFxControllerBa
     private ImageView CitizenIvFailure;
     @FXML
     private Label CitizenLblStatus;
-
-    @FXML
-    private ProgressIndicator IPiProgress;
-    @FXML
-    private ImageView IrisIvSuccess;
-    @FXML
-    private ImageView IrisIvFailure;
-    @FXML
-    private Label IrisLblStatus;
 
     @FXML
     private Button btnRetry;
@@ -79,15 +60,16 @@ public class RegisteringCitizenPaneFxController extends WizardStepFxControllerBa
             }
             else {
                 CitizenLblStatus.setText(resources.getString("label.failedToSendCitizenInfo"));
-                CPiProgress.setVisible(false);
-                CitizenIvFailure.setVisible(true);
-                btnStartOver.setVisible(true);
-                btnRetry.setVisible(!disableRetryButtonForever);
+                GuiUtils.showNode(CPiProgress, false);
+                GuiUtils.showNode(CitizenIvFailure, true);
+                GuiUtils.showNode(btnStartOver, true);
+                GuiUtils.showNode(btnRetry, !disableRetryButtonForever);
+
             }
         }
         else if (request == Request.CHECK_CITIZEN_REGISTRATION) {
             if (successfulResponse) {
-                if (citizenRegistrationStatus == CheckCitizenRegistrationWorkflowTask.Status.PENDING) {
+                if (citizenRegistrationStatus == Status.PENDING) {
                     Context.getExecutorService().submit(() -> {
                         try {
                             int seconds = Integer.parseInt(Context.getConfigManager().getProperty(
@@ -101,92 +83,35 @@ public class RegisteringCitizenPaneFxController extends WizardStepFxControllerBa
                     });
 
                 }
-                else if (citizenRegistrationStatus == CheckCitizenRegistrationWorkflowTask.Status.SUCCESS) {
+                else if (citizenRegistrationStatus == Status.SUCCESS) {
                     CitizenLblStatus.setText(resources.getString("label.successCitizenRegistration"));
-                    CPiProgress.setVisible(false);
-                    CitizenIvSuccess.setVisible(true);
-
-
-                    if (skipIris) {
-                        btnStartOver.setVisible(true);
-                    }
-                    else {
-                        IPiProgress.setVisible(true);
-                        IrisLblStatus.setVisible(true);
-                        //  IrisLblStatus.setText(resources.getString("label.submittingIris"));
-                        request = Request.SUBMIT_IRIS_REGISTRATION;
-                        continueWorkflow();
-                    }
+                    GuiUtils.showNode(CPiProgress, false);
+                    GuiUtils.showNode(CitizenIvSuccess, true);
+                    GuiUtils.showNode(btnStartOver, true);
                 }
 
-                else if (citizenRegistrationStatus == CheckCitizenRegistrationWorkflowTask.Status.HIT) {
+                else if (citizenRegistrationStatus == Status.HIT) {
                     CitizenLblStatus.setText(resources.getString("label.successCitizenRegistrationWithHitResponse"));
-                    CPiProgress.setVisible(false);
-                    CitizenIvFailure.setVisible(true);
-                    btnStartOver.setVisible(true);
-                    btnRetry.setVisible(!disableRetryButtonForever);
+                    GuiUtils.showNode(CPiProgress, false);
+                    GuiUtils.showNode(CitizenIvSuccess, true);
+                    GuiUtils.showNode(btnStartOver, true);
                 }
-                else if (citizenRegistrationStatus == CheckCitizenRegistrationWorkflowTask.Status.ERROR) {
+                else if (citizenRegistrationStatus == Status.ERROR) {
                     CitizenLblStatus.setText(resources.getString("label.failedToRegisterCitizen"));
-                    CPiProgress.setVisible(false);
-                    CitizenIvFailure.setVisible(true);
-                    btnStartOver.setVisible(true);
-                    btnRetry.setVisible(!disableRetryButtonForever);
+                    GuiUtils.showNode(CPiProgress, false);
+                    GuiUtils.showNode(CitizenIvFailure, true);
+                    GuiUtils.showNode(btnStartOver, true);
+                    GuiUtils.showNode(btnRetry, !disableRetryButtonForever);
                 }
 
 
             }
             else {
                 CitizenLblStatus.setText(resources.getString("label.failedToRegisterCitizen"));
-                CPiProgress.setVisible(false);
-                CitizenIvFailure.setVisible(true);
-                btnStartOver.setVisible(true);
-                btnRetry.setVisible(!disableRetryButtonForever);
-            }
-        }
-        else if (request == Request.SUBMIT_IRIS_REGISTRATION) {
-            if (successfulResponse) {
-                IrisLblStatus.setText(resources.getString("label.waitingIrisRegistration"));
-                request = Request.CHECK_IRIS_REGISTRATION;
-                continueWorkflow();
-            }
-            else {
-                IrisLblStatus.setText(resources.getString("label.failedToSendIris"));
-                IPiProgress.setVisible(false);
-                IrisIvFailure.setVisible(true);
-                btnStartOver.setVisible(true);
-                btnRetry.setVisible(!disableRetryButtonForever);
-            }
-        }
-        else if (request == Request.CHECK_IRIS_REGISTRATION) {
-            if (successfulResponse) {
-                if (irisRegistrationStatus == CheckIrisRegistrationWorkflowTask.Status.PENDING) {
-                    Context.getExecutorService().submit(() -> {
-                        try {
-                            int seconds = Integer.parseInt(
-                                    Context.getConfigManager().getProperty("registerIris.inquiry.checkEverySeconds"));
-                            Thread.sleep(seconds * 1000);
-                        } catch (InterruptedException e) {
-                            e.printStackTrace();
-                        }
-
-                        Platform.runLater(this::continueWorkflow);
-                    });
-
-                }
-                else if (irisRegistrationStatus == CheckIrisRegistrationWorkflowTask.Status.SUCCESS) {
-                    IrisLblStatus.setText(resources.getString("label.successIrisRegistration"));
-                    IPiProgress.setVisible(false);
-                    IrisIvSuccess.setVisible(true);
-                    btnStartOver.setVisible(true);
-                }
-            }
-            else {
-                IrisLblStatus.setText(resources.getString("label.failedToRegisterIris"));
-                IPiProgress.setVisible(false);
-                IrisIvFailure.setVisible(true);
-                btnStartOver.setVisible(true);
-                btnRetry.setVisible(!disableRetryButtonForever);
+                GuiUtils.showNode(CPiProgress, false);
+                GuiUtils.showNode(CitizenIvFailure, true);
+                GuiUtils.showNode(btnStartOver, true);
+                GuiUtils.showNode(btnRetry, !disableRetryButtonForever);
             }
         }
     }
@@ -203,31 +128,17 @@ public class RegisteringCitizenPaneFxController extends WizardStepFxControllerBa
 
     @FXML
     private void onRetryButtonClicked(ActionEvent actionEvent) {
-        btnRetry.setVisible(false);
-        btnStartOver.setVisible(false);
-
+        GuiUtils.showNode(btnRetry, false);
+        GuiUtils.showNode(btnStartOver, false);
+        GuiUtils.showNode(CitizenIvFailure, false);
+        GuiUtils.showNode(CPiProgress, true);
 
         if (request == Request.SUBMIT_CITIZEN_REGISTRATION) {
             CitizenLblStatus.setText(resources.getString("label.submitting"));
-            CitizenIvFailure.setVisible(false);
-            CPiProgress.setVisible(true);
         }
         else if (request == Request.CHECK_CITIZEN_REGISTRATION) {
             CitizenLblStatus.setText(resources.getString("label.waitingCitizenRegistration"));
-            CitizenIvFailure.setVisible(false);
-            CPiProgress.setVisible(true);
         }
-        else if (request == Request.SUBMIT_IRIS_REGISTRATION) {
-            IrisLblStatus.setText(resources.getString("label.submittingIris"));
-            IrisIvFailure.setVisible(false);
-            IPiProgress.setVisible(true);
-        }
-        else if (request == Request.CHECK_IRIS_REGISTRATION) {
-            IrisLblStatus.setText(resources.getString("label.waitingIrisRegistration"));
-            IrisIvFailure.setVisible(false);
-            IPiProgress.setVisible(true);
-        }
-
 
         continueWorkflow();
     }
@@ -235,7 +146,7 @@ public class RegisteringCitizenPaneFxController extends WizardStepFxControllerBa
     @FXML
     public void onStartOverButtonClicked(ActionEvent actionEvent)
     {
-        if (citizenRegistrationStatus != null && citizenRegistrationStatus == CheckCitizenRegistrationWorkflowTask.Status.SUCCESS) {
+        if (citizenRegistrationStatus != null && (citizenRegistrationStatus == Status.SUCCESS || citizenRegistrationStatus == Status.HIT)) {
             startOver();
         }
         else {