+++ RELEASE-NOTES.md
@@ -15,6 +15,8 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 - NEW: Register Criminal Clearance Report by Fingerprints.
 - NEW: Inquiry Criminal Clearance Report.
 - NEW: Add live Preview to SingleFingerprintCapturing.
+- CHN: updated citizen enrollment for new iris registration
+- CHN: updated biokit library from v2019.09.1 to v2021.01.1
 
 ---